Refactor schedule parsing and indexing code for simplicity.
